// src/types/data.ts
export default type DataType = {
    compressBoardEnabled: boolean
    compressBoard: number
    compressBlueEnabled: boolean
    compressBlue: number
    compressOrangeEnabled: boolean
    compressOrange: number
    tupferBlueEnabled: boolean
    tupferBlue: number
    tupferOrangeEnabled: boolean
    tupferOrange: number
    coversEnabled: boolean
    covers: number
    waterEnabled: boolean
    waterStart: number
    waterEnd: number
    natcloEnabled: boolean
    natcloStart: number
    natcloEnd: number
    suctionAmniotic: number
    suctionTotal: number
    other: number
    other: number
    estimate: number
    estimate: number
    weight: number
    weight: number
}